Drop response caching on bpod
Caching on bpod was full of assumptions and very hacky, this commits
removes it all. Images are still cached and we prepend the date to
their file name so that it becomes easier to work with them later.
Also they are seperated by a colon so it's easy to remove the date too
with a script or something.

BPOD is not even an official api and it can get closed anyday, bing
uses this to get bpod on bing.com and we are just using it. They still
have pretty pictures though and so I'll still keep it. If users want
to get all information then they can just use -dump flag and save it
to a file.