Change APOD timezone to UTC - 8
Subtract 8 hours from UTC to ensure program doesn't fail. This still
doesn't mean anything, I've emailed them asking about timezone on
server but no response :(

The server returns "400 Bad Request" when you request future date, so
if I request 2020-04-25 on 2020-04-24 23:59 UTC it will return "400
Bad Request" but if I re-request it on 2020-04-25 00:04 UTC it returns
"500 Internal Server Error", I think the API server runs on UTC but
the program that is responsible for syncing the images is running on a
different timezone, which is why it returns "500 Internal Server
Error" instead of "400 Bad Request".

Hopefully this should work, it will work if the program responsible
for syncing images is in or before UTC-8.